+++ src/main/java/com/jds/model/orders/OrderDiscounts.java
@@ -17,7 +17,7 @@ public class OrderDiscounts {
         this.ordersDiscountsList = ordersDiscountsList;
     }
 
-    public int getByDoorId (int door_id){
+    public double getByDoorId (int door_id){
 
         OrderDiscount orderDiscount = ordersDiscountsList.stream()
                 .filter((dis)-> dis.getDoor_id() == (door_id))
